pyproject.toml
[tool.poetry]
name = "notes2notion"
version = "1.0.0"
description = "Import content from Apple Notes into Notion"
authors = ["jheddings <jheddings@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
applescript = "^2021.2.9"
notional = "^0.4.3"
click = "^8.1.3"
pydantic = "^1.10.2"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.10.1"
black = "^22.10.0"
flake8 = "^6.0.0"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
